Change HF1e class
